feat: add interval sealing mode for development node (without relay) (#1533)
# Goal
The goal of this PR is to add interval sealing mode for development node
(no relay chain.)

Closes #1204

### Testing

`./target/debug/frequency --dev --sealing=interval
--sealing-interval=30`

For regression testing:

`./target/debug/frequency --dev --sealing=instant `

`./target/debug/frequency --dev --sealing=manual `

```
USAGE:

      --sealing <SEALING>
          The sealing mode
          
          [default: instant]

          Possible values:
          - instant:
            Author a block immediately upon receiving a transaction into the transaction pool
          - manual:
            Author a block upon receiving an RPC command
          - interval:
            Author blocks at a regular interval specified in seconds

      --sealing-interval <SECONDS>
          The interval in seconds
          
          [default: 12]
   
```
