+++ node/cli/src/cli.rs
@@ -82,11 +82,11 @@ pub struct Cli {
 	pub relay_chain_args: Vec<String>,
 
 	#[cfg(feature = "frequency-no-relay")]
-	#[clap(long = "instant-sealing", help = "deprecated")]
+	#[clap(long = "instant-sealing", help = "Deprecated.  Use --sealing=instant instead.")]
 	pub instant_sealing: bool,
 
 	#[cfg(feature = "frequency-no-relay")]
-	#[clap(long = "manual-sealing", help = "deprecated")]
+	#[clap(long = "manual-sealing", help = "Deprecated.  Use --sealing=manual instead.")]
 	pub manual_sealing: bool,
 
 	/// Instant block sealing
@@ -98,17 +98,8 @@ pub struct Cli {
 
 	/// Interval in seconds for interval sealing.
 	#[cfg(feature = "frequency-no-relay")]
-	#[clap(long, help = "The interval in seconds", default_value = "120", value_name = "SECONDS")]
+	#[clap(long, help = "The interval in seconds", default_value = "12", value_name = "SECONDS")]
 	pub sealing_interval: NonZeroU16,
-
-	/// Whether to create empty blocks in manual and interval sealing modes.
-	#[cfg(feature = "frequency-no-relay")]
-	#[clap(
-		long,
-		default_value = "true",
-		help = "Allows empty blocks to be created when sealing in manual and interval modes"
-	)]
-	pub sealing_allow_empty_blocks: bool,
 }
 
 #[derive(Debug)]
