+++ node/cli/src/run_as_localchain.rs
@@ -1,16 +1,9 @@
 use crate::cli::Cli;
-use cli_opt::SealingMode;
+use frequency_service::block_sealing::frequency_dev_sealing;
 use sc_cli::SubstrateCli;
 use std::process;
 
 pub fn run_as_localchain(cli: Cli) -> sc_service::Result<(), sc_cli::Error> {
-	// Temporary error message
-	if cli.sealing == SealingMode::Interval {
-		return Result::<(), _>::Err(sc_cli::Error::Input(String::from(
-			"Interval sealing is not yet implemented.",
-		)))
-	}
-
 	let runner = cli.create_runner(&cli.run.normalize())?;
 
 	runner.run_node_until_exit(|config| async move {
@@ -25,6 +18,7 @@ pub fn run_as_localchain(cli: Cli) -> sc_service::Result<(), sc_cli::Error> {
 			);
 			process::exit(1);
 		}
-		frequency_service::service::frequency_dev_sealing(config, cli.sealing).map_err(Into::into)
+		frequency_dev_sealing(config, cli.sealing, u16::from(cli.sealing_interval))
+			.map_err(Into::into)
 	})
 }
