+++ node/service/src/lib.rs
@@ -3,6 +3,9 @@
 #![warn(missing_docs)]
 #![warn(unused_extern_crates)]
 
+/// Block sealing
+#[cfg(feature = "frequency-no-relay")]
+pub mod block_sealing;
 pub mod chain_spec;
 pub mod rpc;
 pub mod service;
