+++ scripts/init.sh
@@ -99,6 +99,36 @@ start-frequency-instant)
     --tmp
   ;;
 
+start-frequency-interval)
+  printf "\nBuilding Frequency with runtime interval sealing ...\n"
+  cargo build --features frequency-no-relay
+
+  parachain_dir=$base_dir/parachain/${para_id}
+  mkdir -p $parachain_dir;
+
+  if [ "$2" == "purge" ]; then
+    echo "purging parachain..."
+    rm -rf $parachain_dir
+  fi
+
+  ./target/debug/frequency \
+    --dev \
+    -lruntime=debug \
+    --sealing=interval \
+    --wasm-execution=compiled \
+    --execution=wasm \
+    --no-telemetry \
+    --no-prometheus \
+    --port $((30333)) \
+    --rpc-port $((9933)) \
+    --ws-port $((9944)) \
+    --rpc-external \
+    --rpc-cors all \
+    --ws-external \
+    --rpc-methods=Unsafe \
+    --tmp
+  ;;
+
 start-frequency-native)
   printf "\nBuilding Frequency with runtime instant sealing, native execution ...\n"
   cargo build --features frequency-no-relay
